Function to insert a table into the connected document

Inserts a table at the current position of the view cursor.

rlo_table(x, captiontext, header = "colnames", group_header = NULL,
  common_header = NULL, group_sizes = NULL, footer = NULL,
  factors = NULL, merge_index = NULL, numbered = TRUE, NA_string = "",
  break_before_caption = FALSE, split = FALSE, repeat_headlines = TRUE,
  charheight = NULL, widths = NULL, warn = FALSE)

Arguments

x

A matrix of character vectors to be inserted as a table. If not a matrix, an attempt is made to turn it into a matrix by as.matrix.

captiontext

The text of the caption

header

The names to be used for the columns of the matrix

group_header

If not NULL, the names of column groups

common_header

If not NULL, the common header of all matrix columns

group_sizes

If group_header is not NULL, a vector holding the sizes of column groups

footer

An optional text to be included as a table footer

factors

An optional named list of character vectors that must describe the rows of the matrix object

merge_index

An optional character vector with the names of the factors for which adjacent cells with identical values should be merged

numbered

Should the caption of the table be numbered?

NA_string

The string used for NA values

break_before_caption

Should a page break be insersted before the caption

split

Should it be allowed to split the table across pages

repeat_headlines

Should the headline(s) be repeated?

charheight

An optional way to specify the character height in table cells

widths

An optional way to specify relative columns widths

warn

Should missing paragraph styles give a warning?